+++ src/Sentry/Scope.cs
@@ -127,7 +127,8 @@ namespace Sentry
         }
 
         /// <inheritdoc />
-        public string? Platform { get; set; }
+        [Obsolete("Platform is always csharp, and should not be set by consuming code. This property will be removed in version 4.")]
+        public string? Platform { get; set; } = Constants.Platform;
 
         /// <inheritdoc />
         public string? Release { get; set; }
@@ -360,7 +361,6 @@ namespace Sentry
             Request.CopyTo(other.Request);
             User.CopyTo(other.User);
 
-            other.Platform ??= Platform;
             other.Release ??= Release;
             other.Environment ??= Environment;
             other.TransactionName ??= TransactionName;