+++ test/Sentry.Tests/Protocol/SentryEventTests.cs
@@ -1,5 +1,3 @@
-using Sentry.Tests.Helpers;
-
 namespace Sentry.Tests.Protocol;
 
 public class SentryEventTests
@@ -81,13 +79,6 @@ public class SentryEventTests
 
             return o;
         });
-    }
-
-    [Fact]
-    public void Ctor_Platform_CSharp()
-    {
-        var evt = new SentryEvent();
-        Assert.Equal(Constants.Platform, evt.Platform);
     }
 
     [Fact]